# Parallelize `UpdateAlg.findUpdates`