This uses `parTraverseFilter` instead of `traverseFilter` for finding
updates in `UpdateAlg`. It makes no difference when versions are
obtained from the cache, but can be a significant speed-up if they
are obtained from Coursier.

Here are the total runtimes for working on the scala-steward repo
itself with an empty and fresh cache compared to the main branch:
```
this branch with an empty cache:
Total time: Steward scala-steward-org/scala-steward: 3s 69ms
Total time: Steward scala-steward-org/scala-steward: 3s 219ms
Total time: Steward scala-steward-org/scala-steward: 3s 16ms

this branch with a fresh cache:
Total time: Steward scala-steward-org/scala-steward: 1s 147ms
Total time: Steward scala-steward-org/scala-steward: 974ms
Total time: Steward scala-steward-org/scala-steward: 1s 46ms

main with an empty cache:
Total time: Steward scala-steward-org/scala-steward: 11s 166ms
Total time: Steward scala-steward-org/scala-steward: 11s 174ms
Total time: Steward scala-steward-org/scala-steward: 10s 799ms

main with a fresh cache:
Total time: Steward scala-steward-org/scala-steward: 1s 113ms
Total time: Steward scala-steward-org/scala-steward: 1s 187ms
Total time: Steward scala-steward-org/scala-steward: 1s 28ms
```
If the cache needs to be refreshed this branch only takes ~3s compared
to ~11s on main.