@fthomas 2022-01-31
Here is a comparison of the total runtimes of a few @scala-steward runs before and after this change:
```
before:
2022-01-27 07:08:38,932 INFO  ──────────── Total time: run: 1h 6m 45s 546ms
2022-01-27 09:29:44,961 INFO  ──────────── Total time: run: 1h 27m 56s 606ms
2022-01-27 11:14:28,550 INFO  ──────────── Total time: run: 1h 12m 34s 954ms
2022-01-27 20:05:50,522 INFO  ──────────── Total time: run: 2h 4m 2s 161ms
2022-01-28 04:56:47,670 INFO  ──────────── Total time: run: 54m 58s 975ms
2022-01-28 07:26:13,993 INFO  ──────────── Total time: run: 1h 24m 23s 110ms
2022-01-28 09:07:44,130 INFO  ──────────── Total time: run: 1h 5m 53s 811ms
2022-01-28 11:08:36,325 INFO  ──────────── Total time: run: 1h 6m 44s 136ms
2022-01-28 15:36:52,274 INFO  ──────────── Total time: run: 1h 35m 152ms
2022-01-28 13:13:43,445 INFO  ──────────── Total time: run: 1h 11m 51s 815ms

after:
2022-01-28 17:02:07,203 INFO  ──────────── Total time: run: 1h 13s 406ms
2022-01-28 19:25:00,181 INFO  ──────────── Total time: run: 1h 23m 15s 373ms
2022-01-30 04:40:49,631 INFO  ──────────── Total time: run: 39m 5s 708ms
2022-01-30 06:45:19,282 INFO  ──────────── Total time: run: 43m 33s 7ms
2022-01-30 09:10:23,494 INFO  ──────────── Total time: run: 1h 8m 33s 612ms
2022-01-30 10:50:42,697 INFO  ──────────── Total time: run: 49m 1s 384ms
2022-01-30 12:44:40,380 INFO  ──────────── Total time: run: 42m 51s 685ms
2022-01-30 14:50:09,223 INFO  ──────────── Total time: run: 48m 17s 740ms
2022-01-30 17:34:59,629 INFO  ──────────── Total time: run: 1h 33m 6s 428ms
2022-01-30 19:08:06,079 INFO  ──────────── Total time: run: 1h 6m 18s 148ms
```
The fastest run before was ~55m and it was the only one under one hour. The fastest run after was ~39m and there were 5 runs under one hour.